Optimized the print_with_magic_color() function.
The function was originally generating a random number from 0,255 and then was checking if that number is form the excluded number and till it finds the number from accepted number range , it kept looping in a while loop. This was replaced by using random.choice([accepted int list]) instead.Added annotations , doc string and removed the bool comparison (if bool_value==True:(bad practice) , if bool_value:(good practice). Removed unnecessary string formatting which makes the code un readable and %string formatting might be deprecated in the near future.